add new cleanFailedJob flow

+++ main.go
@@ -33,34 +33,37 @@ func main() {
 
 			pods := k.GetJobsPod(j)
 
-			for _, p := range pods.Items {
+			if pods != nil {
 
-				if helpers.ShouldGetPodLogs() {
-					tail := helpers.GetLogTail()
-					k.GetPodLogs(p, &tail)
-				}
+				for _, p := range pods.Items {
 
-				if helpers.ShouldDeletePod() {
-					if _, ok := j.Labels["deleted-pod-timestamp"]; !ok {
-						k.LabelJob(j)
-						k.DeletePod(p)
-					} else {
-						logging.Warning("Pod previously deleted!")
+					if helpers.ShouldGetPodLogs() {
+						tail := helpers.GetLogTail()
+						k.GetPodLogs(p, &tail)
+					}
 
-						i, err := strconv.ParseInt(j.Labels["deleted-pod-timestamp"], 10, 64)
-						if err != nil {
-							logging.Error(err.Error())
-						}
-						tm := time.Unix(i, 0)
-
-						if time.Now().Sub(tm.UTC()).Seconds() > activeDeadlineSeconds {
-							logging.Warning("Pod was deleted " + strconv.FormatFloat(time.Now().Sub(tm.UTC()).Seconds(), 'f', 6, 64) + " seconds ago and now exceeds to ActiveDeadlineSeconds of " + strconv.FormatFloat(activeDeadlineSeconds, 'f', 6, 64) + " seconds ago, pod will be deleted again!")
+					if helpers.ShouldDeletePod() {
+						if _, ok := j.Labels["deleted-pod-timestamp"]; !ok {
 							k.LabelJob(j)
 							k.DeletePod(p)
 						} else {
-							logging.Warning("Pod will not be deleted again this time!")
-						}
+							logging.Warning("Pod previously deleted!")
 
+							i, err := strconv.ParseInt(j.Labels["deleted-pod-timestamp"], 10, 64)
+							if err != nil {
+								logging.Error(err.Error())
+							}
+							tm := time.Unix(i, 0)
+
+							if time.Now().Sub(tm.UTC()).Seconds() > activeDeadlineSeconds {
+								logging.Warning("Pod was deleted " + strconv.FormatFloat(time.Now().Sub(tm.UTC()).Seconds(), 'f', 6, 64) + " seconds ago and now exceeds to ActiveDeadlineSeconds of " + strconv.FormatFloat(activeDeadlineSeconds, 'f', 6, 64) + " seconds ago, pod will be deleted again!")
+								k.LabelJob(j)
+								k.DeletePod(p)
+							} else {
+								logging.Warning("Pod will not be deleted again this time!")
+							}
+
+						}
 					}
 				}
